Basic test level and player movement
I made a few prefab world objects with static rigid bodies and placed some of them in the test scene to test player collisions.

I made a player prefab.

I made an InputManager object to handle input (I don't know if thats the way we should do it or not though).

I handled WASD inputs to move the player through its script which is found with "FindObjectOfType" (I don't know if this is good or no).

I gave the Entity a movement_speed variable and made a move function which can be used to move player or enemies depending on their speed.

The player is currently able to move and collide with the world objects.